move d3 extension out
